Split error file table
